% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_popup}
\alias{create_popup}
\title{Create popup}
\usage{
create_popup(session, filebody, lang)
}
\arguments{
\item{session}{Shiny app session.}

\item{filebody}{Species names.}

\item{lang}{Language shortcode.}
}
\description{
Create an information popup in locale.
}
