% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_months}
\alias{get_months}
\title{Get months}
\usage{
get_months(lang = "en", format)
}
\arguments{
\item{lang}{Language shortcode.}

\item{format}{Format.}
}
\description{
Get month names in a given language.
}
