% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species_names}
\alias{get_species_names}
\title{Get species names}
\usage{
get_species_names(lang, sp_data)
}
\arguments{
\item{lang}{Language shortcode.}

\item{sp_data}{Species names.}
}
\description{
Get species common names in a locale.
}
