% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Get timestamp}
\usage{
get_timestamp(x)
}
\arguments{
\item{x}{Date.}
}
\description{
Return numeric timestamp value based on a Date. Needed for highcharts
x-axis settings and plotting.
}
