% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_description}
\alias{parse_description}
\title{Parse description}
\usage{
parse_description(text, lang)
}
\arguments{
\item{text}{Description text.}

\item{lang}{Language shortcode.}
}
\description{
Parse description text.
}
