% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_observations}
\alias{tile_observations}
\title{Calculate a tiled average (non-overlapping windows) for yearly observations.}
\usage{
tile_observations(x, value)
}
\arguments{
\item{x}{A data.frame.}

\item{value}{String name of the variable containing numeric observation
values.}
}
\value{
A data.frame with dates and corresponding (rounded) averages.
}
\description{
An average is calculated over a specified window and assigned to the day in
the middle of the window.
}
