% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_value}
\alias{get_value}
\title{Get record value}
\usage{
get_value(season, type, value, records, i18n)
}
\arguments{
\item{season}{Character. "Spring" or "Autumn".}

\item{type}{Character. Migration "Migr" or "Local".}

\item{value}{Character. "date_string" or "Sum".}

\item{records}{Records table.}

\item{i18n}{Translation function.}
}
\description{
Get record value for a species.
}
