% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tile_observations}
\alias{tile_observations}
\title{Calculate a tiled average (non-overlapping windows) for yearly observations.}
\usage{
tile_observations(x, day, value, size)
}
\arguments{
\item{x}{A tsibble.}

\item{day}{String name of the variable containing datetime values.}

\item{value}{String name of the variable containing numeric observation
values.}

\item{size}{Numeric size of the window over which average is calculated.
Must be an odd value.}
}
\value{
A data.frame with dates and corresponding (rounded) averages.
}
\description{
An average is calculated over a specified window and assigned to the day in
the middle of the windows. For example, if size is 5, then the average is
assigned to day at index 3 etc.
}
